Walking
Allows the robot to walk short distances from voice commands. Currently the only options to walk are Forwards, Backwards, Left, and Right. The robot will only move .2 meters in any directions, due to the limited space available on the table it resides on.
To have the robot perform any of these tasks, say either "Move <direction>" or "Walk <direction>"