First finished version of variable walking
This is the first working version of the variable walking program. It will ask for a direction (forwards, backwards, left, right), then a unit of measure (feet, inches), then how far it will be walking (1-12 units).
Currently, it is having issues with recognizing words for the units, but this is most likely due to the nature of its native speech recognition.