Preliminary Red Ball Module
Super early work on the red ball module. Not sure if it would be better to create a Python box from scratch to implement this, or use the built in red ball box. Before continuing on any work with this, we need to get a new red ball that is bigger and has no other colors on it. The current one we have is not very reliable for testing purposes.