Age Guesser Module
I think this is ready to integrate. Getting NAO to recognize your face occasionally doesn't seem to work (or just takes a super long time), but i've only tried the module by running it through Choregraphe while NAO is hooked up. We'll have to do testing when it's actually integrated onto the robot to see the performance of the module.