Expression Guesser
The code/choregraphe workbench looks to all be correct - however this module does not seem to be super reliable as of now. Like the age checker module, i've only tested this through choregraphe while NAO is attached to the computer, so I think it would be best to integrate it and test it autonomously to check the reliability. I've put it in the "InProgress" folder for now in case we make any changes.