Work in Progress Documentation
I've started creating the official documentation / user manual document for our project. My goal with this is to make it more in-line with the styling and format of our SRS and Midterm Progress Report. So far I've created a title page, table of contents, and the section headers. I also copy/pasted some of what Hailey has written in the other documentation document in this folder, but aside from that there is not a lot of actual content here yet. I'll keep working on this through the rest of the week/weekend. Anyone else please feel free to contribute if you'd like!