Completed Mood/Expression Guesser
Mood guesser should be working and ready for integration now, i've been testing it on the physical robot. Similar to AgeGuesser, it's not the most accurate thing in the world, but it's sure to get some good laughs from people who demo the robot.

Also, I renamed the module from ExpressionGuesser to MoodGuesser.